+++ src/main/kotlin/com/goofy/realtime/common/dto/RealtimePageRequest.kt
@@ -1,4 +1,4 @@
-package com.goofy.realtime.common.model
+package com.goofy.realtime.common.dto
 
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
